+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.2] - TBD
+### Changed
+- Improved folding logic of downloading section when only a single artifact is downloaded
+
 ## [0.1.1] - 2022-11-14
 ### Changed
 - Update Future Enhancements section of README